refactor: split the worker and local-server libs (#278)
