+++ local-server/Cargo.toml
@@ -15,6 +15,7 @@ http = { workspace = true }
 hickory-server = { version = "0.25.1", features = ["resolver"] }
 hyper = { version = "1.5.2", features = ["server"] }
 hyper-rustls = { version = "0.27.5", default-features = false, features = [
+    "http1",
     "http2",
     "ring",
 ] }
