+++ worker/src/handlers/mod.rs
@@ -1,0 +1,6 @@
+pub mod proxy;
+pub mod v1;
+
+pub async fn always_ok() -> &'static str {
+    "OK"
+}
