+++ worker/src/tunnel.rs
@@ -1,6 +1,16 @@
 use std::fmt::Display;
 
-use crate::TunnelData;
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize)]
+pub struct TunnelData {
+    pub account_id: String,
+    pub name: String,
+    pub url: String,
+    pub id: String,
+    pub secret: String,
+    pub last_started: u64,
+}
 
 #[derive(Debug)]
 pub enum CreateTunnelError {
